mm/damon/core: add eliglble_report() ops callback
Not every reported access information will be eligible for all DAMON ops
and target.  For example, virtual address access report will be not
eligible for 'padr' ops, or monitoring targets for a process that
different from the process for the repor.  If it is for monitoring
accesses from specific CPU or write, reports from other CPUs or reads
should be ignored.  Add operations set callback for this report
eligibility checking.

Signed-off-by: SeongJae Park <sj@kernel.org>