mm/damon/vaddr: cleanup using pmd_trans_huge_lock()
Three pmd walk functions in vaddr.c are doing pmd_trans_huge() check and
then pmd_lock().  Replace it with pmd_trans_huge_lock(), which does the
check and locking at once.

Yet another advantage of the new code is the simpler handling of pmd
migration entries case.  pmd_trans_huge() doesn't accept migration
entries while pmd_trans_huge_lock() accepts those.  Hence if the pmd is
a migration entry, this change makes the handling of the pmd be finished
within the THP handling part.  The old code requires the handling
further goes down to PTE handling part below.

The old complicated code flow actually contributed to a mistake that
resulted in a softlockup [1] issue.  And this change is suggested by
Hugh on his reply to the issue report, for not only fixing the issue but
also for cleanup of the code.  We decided to make a simpler fix for the
ease of the stable kernels porting.  But as I also acknowledged on the
thread, this change is nice cleanup for both length of the code and the
simplicity of the code execution flow.

[1] https://lore.kernel.org/296c2b3f-6748-158f-b85d-2952165c0588@google.com

Suggested-by: Hugh Dickins <hughd@google.com>
Signed-off-by: SeongJae Park <sj@kernel.org>