selftests/damon/wss_estimation: increase allowed error rate
The test's max error rate was set only with gut feeling on the test
machine that I was using before.  On a smaller machine I use nowadays,
the test fails since the working set is found slower.  Increase the
allowed error rate by increasing the percentile to evaluate from p50 to
p75.

Signed-off-by: SeongJae Park <sj@kernel.org>